[CLI] Add cli commands to hold version (for now) and others (later).
