+++ cli/src/pcluster/cli/entrypoint.py
@@ -30,6 +30,7 @@ import pcluster.api.controllers.cluster_instances_controller
 import pcluster.api.controllers.cluster_operations_controller
 import pcluster.api.controllers.image_operations_controller
 import pcluster.api.errors
+import pcluster.cli.commands as cli_commands
 import pcluster.cli.commands.cluster as cluster_commands
 import pcluster.cli.commands.image as image_commands
 import pcluster.cli.logging as pcluster_logging
@@ -204,7 +205,9 @@ def add_cli_commands(parser_map):
     subparsers = parser_map["subparser"]
 
     # add all non-api commands via introspection
-    for _name, obj in inspect.getmembers(cluster_commands) + inspect.getmembers(image_commands):
+    for _name, obj in (
+        inspect.getmembers(cluster_commands) + inspect.getmembers(image_commands) + inspect.getmembers(cli_commands)
+    ):
         if inspect.isclass(obj) and issubclass(obj, CliCommand) and not inspect.isabstract(obj):
             obj(subparsers)
